refactor(uzicli): split admin verbs into client_admin.go/fake_admin.go (PRD #1017 M2)
Pure motion: move the 9 admin HTTPClient verbs and their FakeClient twins into
per-domain files. go doc -u -all byte-identical; no test edits.
